Fixed seek/read
VFS:
Had to disable input buffering, as the current implementation was really
bad it seemed. Now the output buffering is working pretty well, and I
actually think I'm going to implement the input buffering in userspace
to speed it up as well.

MFS:
Both reading and writing is SLOW, but works. This is mainly due to the
reason that if the file is allocated into ONE HUGE bucket, then it loads
the entire file every time it just wants to read a single byte. So i
have to calculate which bucket I need to read before hand instead, so I
at max load only a bucket, and not the entire file

+++ kernel/System/Syscalls.c
@@ -432,7 +432,7 @@ int ScVfsWrite(FILE *cData, uint8_t *Buffer, size_t Length)
 }
 
 /* Seek File */
-int ScVfsSeek(FILE *cData, size_t Position)
+int ScVfsSeek(FILE *cData, off_t Position)
 {
 	/* Sanity */
 	if (cData == NULL || cData->_handle == NULL)
